changes for using angular material design

+++ Galeria_API/Core/Model/ValidationAttributeValueBetween.cs
@@ -9,7 +9,7 @@ namespace Galeria_API.Core.Model
         public int Max { get; set; } = 10;
         protected override ValidationResult IsValid(object value, ValidationContext validationContext)
         {
-            if (value is int intValue) 
+            if (value is byte intValue) 
             {
                 return intValue >= Min && intValue <= Max
                         ? ValidationResult.Success
